djangoapp/requirements.txt
Django>=5.0
psycopg[binary]
django-environ
environ
boto3
minio
django-storages
python-dateutil
djangorestframework
